requirements.txt
attrs==22.2.0
pip==23.0.1
pytest==7.2.1
selenium==4.8.2
setuptools==67.4.0
pluggy==1.0.0
exceptiongroup==1.1.0
iniconfig==2.0.0
certifi==2022.12.7
urllib3==1.26.14
allure-pytest==2.12.0
allure-python-commons==2.12.0
six==1.16.0